+++ src/backends/sentry_backend_breakpad.cpp
@@ -176,8 +176,8 @@ breakpad_backend_callback(const google_breakpad::MinidumpDescriptor &descriptor,
                 sentry__attachment_free(screenshot);
             }
 
-            if (options->crash_reporter) {
-                sentry__launch_crash_reporter(envelope);
+            if (options->external_crash_reporter) {
+                sentry__launch_external_crash_reporter(envelope);
             } else {
                 // capture the envelope with the disk transport
                 sentry_transport_t *disk_transport