+++ src/backends/sentry_backend_crashpad.cpp
@@ -115,7 +115,7 @@ typedef struct {
     sentry_path_t *event_path;
     sentry_path_t *breadcrumb1_path;
     sentry_path_t *breadcrumb2_path;
-    sentry_path_t *report_path;
+    sentry_path_t *external_report_path;
     size_t num_breadcrumbs;
     std::atomic<bool> crashed;
     std::atomic<bool> scope_flush;
@@ -227,10 +227,7 @@ flush_external_crash_report(const sentry_path_t *report_path,
         sentry__envelope_add_session(envelope, options->session);
     }
 
-    if (sentry__path_create_dir_all(options->run->report_path) != 0
-        || sentry_envelope_write_to_path(envelope, report_path) != 0) {
-        SENTRY_WARN("flushing external crash report failed");
-    }
+    sentry__run_write_external(options->run, envelope);
     sentry_envelope_free(envelope);
 }
 
@@ -267,9 +264,9 @@ crashpad_backend_flush_scope(
         event, "level", sentry__value_new_level(SENTRY_LEVEL_FATAL));
 
     flush_scope_to_event(data->event_path, options, event);
-    if (data->report_path) {
+    if (data->external_report_path) {
         flush_external_crash_report(
-            data->report_path, options, &data->crash_event_id);
+            data->external_report_path, options, &data->crash_event_id);
     }
     data->scope_flush.store(false, std::memory_order_release);
 #endif
@@ -292,11 +289,11 @@ flush_scope_from_handler(
         expected = false;
     }
 
-    // now we are the sole flusher and can flush into the crash report/event
+    // now we are the sole flusher and can flush into the crash event
     flush_scope_to_event(state->event_path, options, crash_event);
-    if (state->report_path) {
+    if (state->external_report_path) {
         flush_external_crash_report(
-            state->report_path, options, &state->crash_event_id);
+            state->external_report_path, options, &state->crash_event_id);
     }
 }
 
@@ -497,15 +494,15 @@ crashpad_backend_startup(
 
     base::FilePath crash_reporter;
     base::FilePath crash_envelope;
-    if (options->crash_reporter) {
+    if (options->external_crash_reporter) {
         char *filename
             = sentry__uuid_as_filename(&data->crash_event_id, ".envelope");
-        data->report_path
-            = sentry__path_join_str(options->run->report_path, filename);
+        data->external_report_path
+            = sentry__path_join_str(options->run->external_path, filename);
         sentry_free(filename);
 
-        crash_reporter = base::FilePath(options->crash_reporter->path);
-        crash_envelope = base::FilePath(data->report_path->path);
+        crash_reporter = base::FilePath(options->external_crash_reporter->path);
+        crash_envelope = base::FilePath(data->external_report_path->path);
     }
 
     std::vector<std::string> arguments { "--no-rate-limit" };
@@ -648,7 +645,7 @@ crashpad_backend_free(sentry_backend_t *backend)
     sentry__path_free(data->event_path);
     sentry__path_free(data->breadcrumb1_path);
     sentry__path_free(data->breadcrumb2_path);
-    sentry__path_free(data->report_path);
+    sentry__path_free(data->external_report_path);
     sentry_free(data);
 }