+++ src/sentry_database.h
@@ -10,7 +10,7 @@ typedef struct sentry_run_s {
     sentry_uuid_t uuid;
     sentry_path_t *run_path;
     sentry_path_t *session_path;
-    sentry_path_t *report_path;
+    sentry_path_t *external_path;
     sentry_filelock_t *lock;
 } sentry_run_t;
 
@@ -45,9 +45,9 @@ bool sentry__run_write_envelope(
 /**
  * This will serialize and write the given envelope to disk into a file named
  * like so:
- * `<database>/reports/<event-uuid>.envelope`
+ * `<database>/external/<event-uuid>.envelope`
  */
-sentry_path_t *sentry__run_write_report(
+sentry_path_t *sentry__run_write_external(
     const sentry_run_t *run, const sentry_envelope_t *envelope);
 
 /**