+++ src/sentry_options.c
@@ -98,7 +98,7 @@ sentry_options_free(sentry_options_t *opts)
     sentry_free(opts->transport_thread_name);
     sentry__path_free(opts->database_path);
     sentry__path_free(opts->handler_path);
-    sentry__path_free(opts->crash_reporter);
+    sentry__path_free(opts->external_crash_reporter);
     sentry_transport_free(opts->transport);
     sentry__backend_free(opts->backend);
     sentry__attachments_free(opts->attachments);
@@ -553,18 +553,19 @@ sentry_options_set_database_path_n(
 }
 
 void
-sentry_options_set_crash_reporter_path(sentry_options_t *opts, const char *path)
+sentry_options_set_external_crash_reporter_path(
+    sentry_options_t *opts, const char *path)
 {
-    sentry__path_free(opts->crash_reporter);
-    opts->crash_reporter = sentry__path_from_str(path);
+    sentry__path_free(opts->external_crash_reporter);
+    opts->external_crash_reporter = sentry__path_from_str(path);
 }
 
 void
-sentry_options_set_crash_reporter_path_n(
+sentry_options_set_external_crash_reporter_path_n(
     sentry_options_t *opts, const char *path, size_t path_len)
 {
-    sentry__path_free(opts->crash_reporter);
-    opts->crash_reporter = sentry__path_from_str_n(path, path_len);
+    sentry__path_free(opts->external_crash_reporter);
+    opts->external_crash_reporter = sentry__path_from_str_n(path, path_len);
 }
 
 #ifdef SENTRY_PLATFORM_WINDOWS
@@ -614,7 +615,7 @@ sentry_options_set_handler_pathw(sentry_options_t *opts, const wchar_t *path)
 }
 
 void
-sentry_options_set_crash_reporter_pathw_n(
+sentry_options_set_external_crash_reporter_pathw_n(
     sentry_options_t *opts, const wchar_t *path, size_t path_len)
 {
     sentry__path_free(opts->crash_reporter);
@@ -622,11 +623,11 @@ sentry_options_set_crash_reporter_pathw_n(
 }
 
 void
-sentry_options_set_crash_reporter_pathw(
+sentry_options_set_external_crash_reporter_pathw(
     sentry_options_t *opts, const wchar_t *path)
 {
     size_t path_len = path ? wcslen(path) : 0;
-    sentry_options_set_crash_reporter_pathw_n(opts, path, path_len);
+    sentry_options_set_external_crash_reporter_pathw_n(opts, path, path_len);
 }
 
 void