+++ src/transports/sentry_disk_transport.c
@@ -28,19 +28,19 @@ sentry_new_disk_transport(const sentry_run_t *run)
 }
 
 static void
-send_report_disk_transport(sentry_envelope_t *envelope, void *state)
+send_external_disk_transport(sentry_envelope_t *envelope, void *state)
 {
     const sentry_run_t *run = state;
 
-    sentry__run_write_report(run, envelope);
+    sentry__run_write_external(run, envelope);
     sentry_envelope_free(envelope);
 }
 
 sentry_transport_t *
-sentry_new_report_disk_transport(const sentry_run_t *run)
+sentry_new_external_disk_transport(const sentry_run_t *run)
 {
     sentry_transport_t *transport
-        = sentry_transport_new(send_report_disk_transport);
+        = sentry_transport_new(send_external_disk_transport);
     if (!transport) {
         return NULL;
     }