+++ src/transports/sentry_disk_transport.h
@@ -13,9 +13,9 @@ sentry_transport_t *sentry_new_disk_transport(const sentry_run_t *run);
 
 /**
  * This creates a new transport that serializes envelopes to disk in the given
- * `reports` directory.
- * See `sentry__run_write_report`.
+ * `external` directory.
+ * See `sentry__run_write_external`.
  */
-sentry_transport_t *sentry_new_report_disk_transport(const sentry_run_t *run);
+sentry_transport_t *sentry_new_external_disk_transport(const sentry_run_t *run);
 
 #endif